pyproject.toml
[project]
name = "ref-schema"
version = "0.1.0"
description = "An RDF schema for capturing references to other protocols and systems"
authors = ["Gabe Fierro <gtfierro@mines.edu>"]
license = "bsd-3-clause"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "pyontoenv>=0.4.0a10",
    "pyshacl>=0.30.1",
    "pytest>=8.3.5",
    "rdflib>=7.1.4",
]
